Move fetching new cards to library service
